requirements.txt
async-timeout==5.0.1
certifi==2025.1.31
charset-normalizer==3.4.1
filelock==3.16.1
fsspec==2025.3.0
hiredis==3.1.0
idna==3.10
Jinja2==3.1.6
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
opencv-python==4.8.1.78
Pillow==10.1.0
python-decouple==3.8
redis==5.2.1
requests==2.32.3
scikit-learn==1.3.2
scipy==1.10.1
sympy==1.13.3
threadpoolctl==3.5.0
torch==2.1.2
torchvision==0.16.2
triton==2.1.0
typing_extensions==4.13.2
urllib3==2.2.3
